transcriptionService/requirements.txt
setuptools
setuptools-rust
pyaudio
SpeechRecognition
torch
numpy
fastapi[standard]
openai-whisper
librosa
edge-tts